requirements.txt
numpy==1.19.1
pandas==1.1.0
mysql-connector-python==8.0.12
SQLAlchemy==1.3.18
pymongo==3.11.0
google-api-python-client==1.10.0
oauth2client==4.1.3
inflection==0.5.0
requests==2.24.0